# feat: comprehensive code optimization and documentation enhancement