--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,4 @@
 .env
 
 trace_test.env
+.DS_Store