## Summary

This PR includes comprehensive code optimization, test coverage improvement, and documentation enhancement for the trace library.

### Key Improvements

**Code Quality & Testing:**
- Improved test coverage from 38.1% to 90.7% (89/89 tests passing)
- Added 5 comprehensive test suites with 50+ test cases
- Created 19 performance benchmark tests
- Fixed critical bugs including SetSpanExtractor comparison logic
- Removed 20+ unused functions reducing maintenance burden

**Performance Optimization:**
- Consolidated severity methods for better performance
- Added buffer reuse in TracerTagBuilder for memory efficiency
- Achieved 3M+ operations/sec for severity operations
- No-op span optimization provides 14x performance improvement when tracing disabled

**Documentation Enhancement:**
- Added comprehensive performance analysis to both README files
- Included benchmark results and test coverage statistics
- Optimized markdown formatting for professional appearance
- Removed emojis for clean, enterprise-ready documentation
- Updated both English and Chinese documentation with consistent formatting

**Technical Fixes:**
- Fixed SetSpanExtractor slice comparison issue
- Renamed misspelled file (varsBulder.go → varsBuilder.go → deleted as unused)
- Fixed integration test expectations for OTLP behavior
- Resolved all markdown linting issues

### Test Coverage Improvement

- **Unit Tests**: Core functionality complete testing
- **Integration Tests**: End-to-end cross-service tracing verification
- **Concurrent Tests**: Multi-threaded safety verification  
- **Performance Tests**: Critical path performance benchmarks
- **Edge Case Tests**: Error handling and boundary conditions

### Performance Benchmarks

```
BenchmarkSeveritySpan_Debug         3,250,894 ops/sec    398.8 ns/op    912 B/op
BenchmarkSeveritySpan_Info          3,084,661 ops/sec    410.0 ns/op    912 B/op
BenchmarkNoopSpan_Debug            36,734,693 ops/sec     28.04 ns/op      0 B/op
```

## Test Plan

- [x] All 89 tests pass with 90.7% coverage
- [x] Performance benchmarks show consistent results
- [x] Integration tests with Jaeger successfully verified
- [x] Markdown linting passes without errors
- [x] No breaking changes to public API

🤖 Generated with [Claude Code](https://claude.ai/code)